[MISC] move header of sam_file
API Header to reflect content better:
* #include <seqan3/io/sam_file/{misc => sam_flag}.hpp>

Detail Header:
* #include <seqan3/io/sam_file/detail{ => /cigar}.hpp>
* #include <seqan3/io/sam_file/{ => detail}/format_sam_base.hpp>
