+++ include/seqan3/io/sam_file/sam_flag.hpp
