+++ test/snippet/io/sam_file/get_cigar_vector.cpp
@@ -1,7 +1,7 @@
 #include <seqan3/alphabet/gap/gapped.hpp>
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/core/debug_stream.hpp>
-#include <seqan3/io/sam_file/detail.hpp>
+#include <seqan3/io/sam_file/detail/cigar.hpp>
 
 int main()
 {
